requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
black==22.10.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.1.2
django-appconf==1.0.5
django-countries==7.4.2
django-debug-toolbar==3.7.0
django-imagekit==4.1.0
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
mailchimp-marketing==3.0.75
MarkupSafe==2.1.1
mypy-extensions==0.4.3
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathspec==0.10.1
pilkit==2.0
Pillow==9.3.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
